DashboardSubsystem class + nonstatic support
