+++ src/main/java/frc/robot/Robot.java
@@ -5,7 +5,7 @@
 package frc.robot;
 
 import frc.robot.common.LocalADStarAK;
-import frc.robot.common.components.DashboardAutoUpdater;
+import frc.robot.common.components.dashboard.DashboardAutoUpdater;
 import frc.robot.common.components.RobotContainerRegistry;
 import frc.robot.common.components.RobotExceptionHandler;
 import frc.robot.common.components.RobotUtils;
