+++ src/main/java/frc/robot/common/subsystems/drive/SwerveDriveSubsystem.java
@@ -8,6 +8,8 @@ import java.util.function.DoubleSupplier;
 import java.util.function.Supplier;
 
 import frc.robot.CommonConstants;
+import frc.robot.common.annotations.DashboardVariable;
+import frc.robot.common.components.dashboard.DashboardSubsystem;
 import frc.robot.pearce.PearceConstants;
 import lombok.Getter;
 import org.apache.commons.math3.analysis.polynomials.PolynomialSplineFunction;
@@ -48,7 +50,6 @@ import edu.wpi.first.units.Units;
 import edu.wpi.first.wpilibj.DriverStation;
 import edu.wpi.first.wpilibj.smartdashboard.Field2d;
 import edu.wpi.first.wpilibj2.command.Command;
-import edu.wpi.first.wpilibj2.command.SubsystemBase;
 import frc.robot.common.components.RobotUtils;
 import frc.robot.common.swerve.RAWRNavX2;
 import frc.robot.common.swerve.RAWRSwerveModule;
@@ -63,7 +64,7 @@ import frc.robot.common.components.hardware.SwerveHardware;
  *
  * @since 2025
  */
-public class SwerveDriveSubsystem extends SubsystemBase implements AutoCloseable {
+public class SwerveDriveSubsystem extends DashboardSubsystem implements AutoCloseable {
 
   /**FOR TESTING ONLY - Skip loading the drive train*/
   public final boolean enableDrive = true;
